+++ app/src/main/java/com/codepath/kpu/pix/models/IGPostComment.java
@@ -4,7 +4,7 @@ package com.codepath.kpu.pix.models;
  * Created by kpu on 2/6/16.
  */
 public class IGPostComment {
-    public double createdTime;
+    public long createdTime;
     public String text;
     public IGUser author;
 }
